refactor: Move base types to proper single file. #32
