+++ src/authorship_attribution/_internals/types/aliases.py
